requirements.txt
django==4.0.5
djangorestframework